container/cdc/analyseur/src/requirements.txt
geographiclib==2.0
geopy==2.3.0
mysql-connector-python==9.2.0
protobuf==3.20.1
